requirement.txt
streamlit
langchain_qdrant
langchain_openai
openai
pypdf
python-dotenv
tiktoken